+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+# Unreleased
+- Only consider the following types as javascript `undefined`, `"text/javascript"` and legacy `"application/javascript"`
+
 # 10.5.4
 - Use `process.nextTick` instead of `setTimeout` for deferring calls to intersection observer callback
 - Never run observer callback on unobserved elements